Add files via upload
JavaScript is a versatile programming language, and one of its fundamental data types is the string. A string in JavaScript is a sequence of characters enclosed within single (' '), double (" "), or backticks (`) quotes. Whether you're manipulating text, building user interfaces, or performing data validation, understanding JavaScript strings is crucial. In this section, we'll delve into the basics and explore some common operations and methods related to strings.